refactor: trip search api 객체 통합
