--- domain/src/main/java/com/tago/domain/trip/dto/TripRecommendDto.java
@@ -1,26 +1,0 @@
-package com.tago.domain.trip.dto;
-
-import com.querydsl.core.annotations.QueryProjection;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TripRecommendDto {
-    private Long tripId;
-    private LocalDateTime dateTime;
-    private String imageUrl;
-    private String name;
-    private int totalTime;
-    private int maxMember;
-    private int currentMember;
-    private List<String> places;
-
-}
